chore: 🔧 organized queries, action functions
